pyproject.toml
[project]
name = "research-automation"
version = "0.1.0"
description = "Video-based health monitoring research automation (Neurological & Cardiovascular)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # CLI
    "typer>=0.15.0",
    "rich>=13.0",
    # Video / CV
    "opencv-python>=4.8",
    # Signal processing
    "scipy>=1.11",
    "numpy>=1.24",
    # Database
    "sqlalchemy>=2.0",
    # Config
    "pyyaml>=6.0",
    # HTTP
    "httpx>=0.27",
    # PDF
    "pymupdf>=1.24",
    # LLM
    "anthropic>=0.40",
    # YouTube
    "yt-dlp>=2024.0",
    # Academic APIs
    "arxiv>=2.0",
    # Data loading
    "pandas>=2.0",
    # HuggingFace (optional dataset download)
    "huggingface_hub>=0.20",
    # Experiment tracking
    "mlflow>=2.10",
    # Templating
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]
cv = [
    # MediaPipe for pose/face detection (not available on all platforms)
    "mediapipe>=0.10",
]
ecg = [
    # ECG signal processing
    "neurokit2>=0.2",
    "wfdb>=4.0",
]
audio = [
    # Audio processing (requires cmake for llvmlite)
    "librosa>=0.10",
]

[project.scripts]
research = "research_automation.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_automation"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
